feat: Send the query method to room service
Ugly hack: I need the JWT when submitting the roomclient method. A potential solution is to keep the JWT in a sort-of global context until the end of the method, something akin to what Spring Boot does (I think?) but I don't wanna mess with that. It's not idiomatic Go to do things implicitly.